refactor webcams
