+++ routes/src/lib.rs
@@ -6,3 +6,4 @@ pub mod printing;
 pub mod spaces;
 pub mod tools;
 pub mod transit;
+pub mod webcams;
