+++ routes/src/webcams.rs
@@ -1,0 +1,6 @@
+use axum::{routing::get, Router};
+use handlers::github::webcams_handler;
+
+pub fn router() -> Router {
+    Router::new().route("/", get(webcams_handler))
+}
