+++ types/src/lib.rs
@@ -3,3 +3,4 @@ pub mod dictionary;
 pub mod food;
 pub mod spaces;
 pub mod transit;
+pub mod webcams;
